refactor(services): refactor: Update services , separate roles
Closes #224
