+++ src/services/users.ts
@@ -1,77 +1,9 @@
-import { RolesUserResponse, UserResponse, UserPayload, UserUpdatePayload, Role } from '../utils/types';
+import { User, UserPayload, UserUpdatePayload } from '../utils/types';
 import axios from "./axios";
 
-export const getRoles = (): Promise<RolesUserResponse> => {
-  return axios
-    .get("/api/roles")
-    .then(function (response) {
-      return response.data;
-    })
-    .catch(function (error) {
-      if (error.response) {
-        throw error.response;
-      }
-      throw error;
-    });
-}
 
-export const getRole = (roleName: string): Promise<Role> => {
-  return axios
-    .get(`/api/roles/${roleName}`)
-    .then(function (response) {
-      return response.data;
-    })
-    .catch(function (error) {
-      if (error.response) {
-        throw error.response;
-      }
-      throw error;
-    });
-}
 
-export const updateRole = (role: Role): Promise<void> => {
-  return axios
-    .put(`/api/roles/${role.Name}`, role)
-    .then(function (response) {
-      return response.data;
-    })
-    .catch(function (error) {
-      if (error.response) {
-        throw error.response;
-      }
-      throw error;
-    });
-}
-
-export const deleteRole = (roleName: string): Promise<void> => {
-  return axios
-    .delete(`/api/roles/${roleName}`)
-    .then(function (response) {
-      return response.data;
-    })
-    .catch(function (error) {
-      if (error.response) {
-        throw error.response;
-      }
-      throw error;
-    });
-}
-
-export const postRoles = (payload: Role): Promise<void> => {
-  return axios
-    .post("/api/roles", payload)
-    .then(function (response) {
-      return response.data;
-    })
-    .catch(function (error) {
-      if (error.response) {
-        throw error.response;
-      }
-      throw error;
-    });
-}
-
-export const getUsers = (): Promise<UserResponse> => {
+export const getUsers = (): Promise<User[]> => {
   return axios
     .get("/api/users")
     .then(function (response) {
@@ -85,7 +17,7 @@ export const getUsers = (): Promise<UserResponse> => {
     });
 };
 
-export const getUserById = (userId: number): Promise<UserResponse> => {
+export const getUserById = (userId: number): Promise<User[]> => {
   return axios
     .get(`/api/users/${userId}`)
     .then(function (response) {
